+++ src/include/taler_bank_service.h
@@ -108,7 +108,7 @@ typedef void
 (*TALER_BANK_AdminAddIncomingResultCallback) (void *cls,
                                               unsigned int http_status,
                                               enum TALER_ErrorCode ec,
-                                              long long unsigned int serial_id,
+                                              uint64_t serial_id,
                                               struct GNUNET_TIME_Absolute timestamp,
                                               const json_t *json);
